+++ src/solids4FoamModels/mechanicalConstitutiveLaw/mechanicalConstitutiveLaws/neoHookeanElasticMisesPlasticMechanicalConstitutiveLaw/neoHookeanElasticMisesPlasticMechanicalConstitutiveLaw.C
@@ -246,7 +246,7 @@ neoHookeanElasticMisesPlasticMechanicalConstitutiveLaw
     {
         FatalIOErrorInFunction(dict)
             << "Invalid Poisson's ratio nu = " << nu_.value()
-            << ". Expected -1 <= nu for linear elasticity."
+            << ". Expected -1 < nu for linear elasticity."
             << exit(FatalIOError);
     }